requirements.txt
flask==2.2.2
colorama==0.4.6
docstring-parser==0.15
python-dotenv
fastapi==0.98.0
kivy==2.2.1
kivymd
openai==0.27.5
pvporcupine==2.2.0
pvrecorder==1.1.1
pydub==0.25.1
starlette==0.27.0
tiktoken==0.3.3
tqdm==4.65.0
uvicorn==0.22.0
websockets==11.0.3
PyYAML
SpeechRecognition
weaviate-client
watchdog
python-multipart
GitPython
geocoder